Move POIROT to rare_variant
